Reference scss file in assets-manifest.json
This is how to load the file generated from application.scss by webpack.
When you reference files via the nunjucks getAssetPath() helper it looks
in the assets-manifest.json file for the key you pass in and then loads
the webpack generated file with a contents hash. This means you don't
have to update this reference everytime the files contents and then the
file hass changes